+++ packages/core/types/kernel.ts
@@ -252,7 +252,7 @@ export type Execution = {
 
 export type KERNEL_V2_VERSION_TYPE = "0.0.2" | "0.2.2" | "0.2.3" | "0.2.4"
 
-export type KERNEL_V3_VERSION_TYPE = "0.3.0" | "0.3.1" | "0.3.2"
+export type KERNEL_V3_VERSION_TYPE = "0.3.0" | "0.3.1" | "0.3.2" | "0.3.3"
 
 export type KERNEL_VERSION_TYPE =
     | KERNEL_V2_VERSION_TYPE